requirements.txt
Django==2.2.7
djangorestframework==3.10.3
mypy==0.740
mypy-extensions==0.4.3
pytz==2019.3
sqlparse==0.3.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
